+++ src/lib/types/util.types.ts
@@ -11,3 +11,10 @@ export type IntersectNonVoid<A, B> =
     ? void : B
     : B extends void
     ? A : A & B
+
+/**
+ * Removes first item from tuple
+ */
+export type SliceFirst<
+    List extends any[]
+> = List extends [infer _Head, ...infer Tail] ? Tail : List;